LOGOS_AGI/runtime/gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
pyyaml==6.0.1
structlog==23.2.0
prometheus-client==0.19.0
opentelemetry-distro==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
slowapi==0.1.9
